restructure/reorganize a couple things
