+++ src/app/center-bottom-content/components-bottom/report-builder/workflow-selection/workflow-selection.component.html
@@ -1,0 +1,3 @@
+<div *ngFor="let workflow of workflows; let i = index">
+    <button (click)="addRemoveWorkflow(workflow)">{{ workflow.title }}</button>
+</div>
