+++ src/app/sidebar-left/sidebar-left.component.html
@@ -91,7 +91,9 @@
 				<div class="v-tab-body">
 
 					<div *ngIf="discoverTab == 'one'">
-						Tab One
+						<div>Testing</div>
+						<h3>testing</h3>
+						<p>Testing even more! I'm testing so much I can't believe it! lakjdkljfkljdflkjd ljskdfjskdfjlsjflkjdkfj </p>
 					</div>
 
 					<div *ngIf="discoverTab == 'two'">
